+++ tor/core/__init__.py
@@ -1,0 +1,5 @@
+import os
+
+__version__ = '0.6.0'
+
+__HEARTBEAT_FILE__ = os.getenv('HEARTBEAT_FILE', 'heartbeat.port')