+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 We follow [Semantic Versioning](http://semver.org/) as a way of measuring stability of an update. This
 means we will never make a backwards-incompatible change within a major version of the project.
 
+## UNRELEASED
+
+- Merges in `tor_core` library again under `tor.core`
+
 ## v3.6.0 (2018-12-2)
 
 - Add handling for `unclaim` comments (credit: @itsthejoker)